Use TraverseOptions when specifying contraction
Instead of specifying only a mode & optimize when contracting graphs,
use TraverseOptions so that more options may be specified.

A TraverseOptions.similarity method is added which rates how close to
one another to TO-s are. Currently only the values of
wheelchairAccessible, mode and optimize are takein into account.

The most similar contracted graph is used, if a minimum similarity
threshold is passed.