add class to normalize transit stop names
this will be used to merge stops into stations, for example where there are four
bus stops at an intersection with similar names, but with the order of the street
names reversed and some abbreviations used in a non-uniform way.

merging stops will simplify profile routing greatly, from an algorithmic and speed
perspective, as well as that of the end user who will see better grouping of
travel options.

of course this prototype is limited to English and tailored to the Arlington project
so it would need to be generalized some day.