remove SampleOperator
Removed SampleOperator and its sole implementation to fetch elapsed time. This abstraction of operators combining an SPT with a SampleSet yielding a ResultSet was useful in reasoning through how Analyst should work, but in reality we always look at travel time because that’s what the search minimizes, so the operator is always the same.