Added preliminary bollard support
Working on #1914.

Currently roads are split if barrier=bollard tag is on node.
And this node becomes BarrierVertex.

During streetTraversal it is checked if start or end vertex are
bollards. If they are CAR isn't allowed to drive on this edge.
TODO: more testing. And bollard acces support. (By default it allows
walking and bicycle)