Add support for long walkstep server localization
Working on #2070
It adds longDescription and shortDescription field to WalkStep.
In longDescription is a long text description of a step:
- Start on [street name] heading [absolute direction]
- Take roundabout [clockwise/counter clockwise] to [ordinal exit
  number] exit on [street name]
- [Relative direction] to continue on [street name]
- [Relative direction] on to [street name]

ShortDescription is currently empty. It will be based on short text on
android OTP which is shown in map markers.

Relative direction are:
- Left, right, hard left, slight right, U-turn left, elevator etc.

Absolute direction are:
- north, south, east, west, northeast etc.

Currently it uses immutable maps for translations of ordinal exit
numbers and relative/absolute directions. All translatable strings are
used with T.tr function call which enables them to be automatically
extracted.
Localization of parameters (ordinal exit numbers, relative and absolute
direction) happens currently in WalkStep.

Message format ("Text {0} other text") is used to add parameters to
strings.

Note: Translations are enabled but don't exist yet in PO files or
Resource Bundles.