adjust function signature, allowing compilation to succeed.
I am not entire sure where this problem came from.
42357ee44f on 12 May, changes only whitespace (removes tabs).
f349cc8463 on 15 June changed a function call in the same section that was reformatted.
For some reason, the earlier whitespace change took priority in the merge, and the function call was wrong.

This one caused a compilation error and was caught, but who knows what else has gone wrong.