remove TransportNetwork packages, which are now in R5 project
Also removed some adapter constructors and methods that connected TransportNetworks to OTP.
This still leaves quite a lot of duplicate code in OTP and R5, because all the Transport Analyst worker and broker code is in both.
We'll eventually remove it from OTP once R5 has been put into production.