Fix test case for intermediate places
Note that in LegSwitchingEdge I had to clear the "last pattern" of the state
editor because otherwise the TransitBoardAlight.traverse() method would 
choke on taking the same pattern (route) again - and return null. However, if
you travel from A to C via B, then there is nothing wrong with taking the
same bus in B that you came with.