Fix for #1077 by removing the platform bonus
The platform bonus is meant to snap to a nearby platform if it is close.
However, this can result in strange endpoint selection where a near road
is ignored in favor of a relatively far platform. Also note that the
platform condition "(edge.getStreetClass() & platform) != 0" was tested
for twice and also got weighted twice: once by dividing and once by
subtracting.
I added a test case which fails in the old behavior but succeeds in with
the code in this branch.