Fix frequency bounds checking
Previously, stop offsets were not taken into account, so you could not
ride on a trip past it's frequency end time, even if you could have
boarded it at an earlier stop (when it was still within the frequency
range). This fix allows that to happen.

This also fixes stop time calculation for exact trips where the (end -
start) is not a multiple of headway. Previously, it calculated based on
the end time, but that's a wrong assumption to make.

Tests for the fixes have also been added.