remove dependency on conveyal jackson2-geojson
this module defines some geojson serializers.
they referred to the old JTS package names which have changed in new versions of JTS.
we could patch com.conveyal.jackson2-geojson and re-release it, however JTS now includes its own GeoJSON writing capability, which we are likely to migrate toward. So as a stopgap measure I've copied in our GeoJSON serializer classes and changed the JTS package names they reference. That original code is under Apache 2 license, and Conveyal authorizes its inclusion and/or relicensing here as needed.
Note that this is only used in a few places in OTP.