Add experimental backward search implementation of the IBH
This was an experiment to see if any improvements could be made by making the backward search also explore pre-transit vertices. After experimenting with a few different variations of this implementation, it does not appear to be better than using a Euclidean heuristic for the pre-transit vertices. It's still exploring probably more vertices than necessary during the backward search.