Do not create on-board vertices and edges when building graph.
OTP2 is now using the Raptor algorithm, so not performing a graph search
on an in-memory graph of objects. It is still copying its raptor data
tables from the existing OTP data tables so there's a lot of duplicate
information in memory, but the in memory graph of vertex and edge
objects now only represents the street network.

This has the added advantage that we can use the existing graph search
algorithms to perform street searches for access and egress to transit,
without those searches ever progressing into the transit network.