remove classes for on-board transit edges and vertices
Note that now if you build a graph with only transit and no OSM, many
vertices will be lost because they are not referenced by any edge.

This breaks large numbers of tests and some real-time update code,
because those are so coupled to the object graph representation of the
network. The tests will need to be completely reworked, hopefully in a
way that does not look at the routing internals, only the responses.
I'm leaving the (broken) tests in for now as reminders of everything
we need to test.