Keep reference to source OTP1 TripTimes in OTP2 Raptor TripSchedules
This allows you see realtime information like:
originalTripTimes.getRealTimeState()
originalTripTimes.isScheduled()
originalTripTimes.getDepartureDelay(boardStopIndexInPattern)
when creating the final itinerary for the end user.

It would also be possible to skip copying the arrival/departure times
into the OTP2 TripSchedules and just read through to the underlying
TripTimes. But this might have a performance penalty since it's used in
a tight inner loop.