Reimplement TripScheduleImpl to avoid copying arrival/departure times.
The arrival/departure times in the original TripTimes are
deduplicated arrays making use of a time offset. By copying them we were
increasing the amount of memory used significantly. The TripTimes are
considered immutable once committed to a TimetableSnapshot so we can
reuse their internal arrays - they should never be changed.

This commit provides two alternate implementations of TripSchedule.
One is a wrapper that just calls the methods of the underlying TripTimes.
The other copies a reference to the arrays in the TripTimes and has its
own equivalent methods to find departure and arrival times. It is
possible that the wrapper implementation is slower due to pointer
chasing. If so, another option is to just make OTP1 TripTimes implement
the OTP2 TripSchedule interface, avoiding any copying or indirection.