Downgrade to surefire 2.19.1
The combination of Java 11, the junit version, the surefire plugin version and
the plugin config `forkCount` and `reuseForks` caused a variety of
failures when running the tests, most frequently a ClassNotFoundException like this:

ClassNotFoundException: org.apache.maven.plugin.surefire.StartupReportConfiguration

It is not entirely clear what the correct way to fix this but it occurs
even when you use version 3.0.0-M3 of the surefire plugin, which is the
newest released version as of writing this. Perhaps also upgrading to
junit 5 might resolve this, but that seemed a little beyond the scope of
the PR.

The least invasive choice was to downgrade the surefire plugin version.

https://github.com/junit-team/junit5/issues/1367
https://jira.apache.org/jira/browse/SUREFIRE-1520
https://issues.apache.org/jira/browse/SUREFIRE-1534
https://stackoverflow.com/questions/53437819/maven-surefire-and-jdk-11