chore: temporarily use transitfeeds.com URL for Massport GTFS (#38)
The version of the GTFS at Massport's URL is set to expire in the
past, meaning that we can't use it to plan any trips. We took the
latest version from transitfeeds.com (appears to be based on the
Trillium version) and uploaded it to a temporary location in S3.