docs: update readthedocs links
The versions are necessarily named after branches or tags so it was impossible to update the v2.0 docs version (named after a nonmoving tag).
For now, docs version "latest" is the default and points to master, which will only receive docs updates until the next release.
The "stable" version doesn't seem to work with our release tags starting with the letter v.