docs: Add OTP2 REST API docs URL to the Dev Guide
We were interested in using the OTP2 REST API, but struggled to find the REST API docs in the current OTP2 documentation. Eventually we found the OTP1 REST API docs URL and plugged in the OTP2 version numbers to stumble on it.

We looked at a lot of the documentation, including this Developer Guide and:
* http://docs.opentripplanner.org/en/dev-2.x/Basic-Tutorial/
* http://docs.opentripplanner.org/en/dev-2.x/OTP2-MigrationGuide/#rest-api
* http://docs.opentripplanner.org/en/dev-2.x/Version-Comparison/

...among others but couldn't find any reference (the links that are there point to the source files on GitHub). 

Apologies if there is already a link to this somewhere else and we just overlooked it - alternate suggestions for this text or a better way to surface these docs for other developers is welcome.