Create separate requests for each sub-request
... so that each has appropriate vertices chosen for its mode,
specifically if the street needs to be CAR accessible.

Because of this the scoping of requests is changed in RoutingWorker so
that it is explicit within a try-with-resources.

For the transit-part of the requests to be cleanly scoped the itinerary
generation is slightly modified so that it works with temporary edges
which were removed from the graph.