fix: NPE in Raptor PathMapper for FLEX
A FLEX access leg might arrive on-board to the first stop processed by Raptor.
Then the journey may continue with a transfer (which is not possible for
regular walking access). The ForwardPathMapper did not support this case. Several
new unit tests cases is added to make sure the mapper handles all possibilities.