fix: Travel back in time, optimize transfers #3491
When posible transfers exist AFTER the stop-place the egress start, then
the TransfersPermutationService will board the trip AFTER the alight
stop position/time. This result in riding the buss back in time. This
is fixed in this commit. For trips with FLEX access the generalized-cost
is also adjusted to account for transfer-cost.