refactor: Calculate board cost
Raptor calculate a onBoardRelativeCost and a transitArrivalCost.
These calculations is done with out reusing any logic. But, when
this become problematic when we want to add a "sophisticated" transfer
calculation - we do not want to do it twice and the transfer information
is lost at the time we calculate the transit stop arrival cost. To solve
this we can calculate the cost at board time and reuse this value when
calculating the onBoardRelativeCost and a transitArrivalCost.