refactor: Move CostCalculator impl out from Raptor
This is a pure refactor and no functionality is changed. The cost
calculation belong to the OTP internal model and should not be part of
the Raptor algorithm implementation. This enable us to switch the cost
calculation and use OTP model information in the cost calculation
without passing the information through Raptor. In the future we would
like to extend Raptor to use 2 or more cost criteria, this commit is a
step in that direction.