bug: In optimize transfer service decorating path
When Raptor return a path with a transfer at another stop than the
optimal stop to transfer, then optimized transfer service will find the
correct stop/path, but the transfer information is not attached to the
new path in a correct way. The OptimizedPath map of transfers(transfersTo)
are updated with the correct transfer, but the key used is the old leg,
not the new. This causes the lookup to fail later.