feat: Link StopPlace with the correct TariffZone(FareZone)
This commit lookup the TariffZone using both the id and version
number, if no version number is given the most current version is used.

The FareZone index will only have the "most current" version of a
FareZone, newer and older versions are not kept - but might be referenced
by Stations; Hence it is possible that a Station reference a FareZone,
witch is not in the index.