refactor: extract TransferConstraint w/facilitated transfer
The TransferConstraint is extracted into a immutable value-object. To do
routing in Raptor the cost calculator only need the value object, while
the path builder and itinerary still need the extended transfer
information, but we should get rid of this - since it is redundant.

Guaranteed and stay-seated transfers are "facilitated" transfers where
we ignore board-, transfer-, and alight-slack. We will also add a grace
period in the future for these transfers.