Parallelize trip planning calculations
OTPFeature.ParallelRouting is added to allow enabling/disabling the
parallelisation, which uses ForkJoinPool.commonPool().

The direct street, direct flex and transit searches are run in
parallel, along with the transit access and egress searches.

DebugTimingAggregator is adjusted to show the per-item and total times
so that the values are correct even if running in parallel.