fix: Cleanup transfer priority cost calculation
There is an error in the logic in OptimizedPathTail line 195 and the
OptimizedPath line 90, when adding the priority cost. In both places a
cost was added for path legs witch did not represent a transfer. Since
the transfer constraint is on the transit-leg BEFORE the optional transfer,
this logic is a bit complicated. This commit fixes this, and refactor
the priority cost calculation, so the cost is always positive.

Previously the transfer priority cost was calculated in the
TransferPriority enum, but part of the information is in the
TransferConstraint, so it make sense to move some of the logic up to he
TransferConstraint.