refactor: Improve the trip-times search
The trip-times-search is used in many places to find boar- and alight-times for a given RaptorTripSchedule. There are several version supporting forward and reverse Raptor search results, but the current version handle the NO_WAIT_WORKER and the 0 minutes transfers badly. To improve this a new search witch find the colosest trip with respect to an approximate time is added, and the existing methods refactored. New unit tests are added. We will use the new version later.