feet: Add support for duplicate access/egress stop
The Raptor routing engine only support one access/egress leg for each stop. The McRaptor do support
this, but not the Standard Raptor or BestTime Raptor. The BestTime will not fail, but pick an
arbitrary access/egress while the Standard will fail. The reason way this has not failed in normal
operation is that the BestTime is used in the Heuristics. But, for transfers with constraints we need
 to use the Standard Raptor state instead.