feet: Add constrained transfer id
In NeTEx an interchange have an id, in GTFS a transfer do not. We include it here to enable
debugging, logging and system integration. Note! OTP do not use the id, and it is just passed
through OTP. There is no service in OTP to look up a transfer by the id.