fix: Use correct mode when removing alight-slack for constrained transfers
The SlackProvider requires the current trip pattern to be set before retrieving slack. If the
setCurrentPattern method is not called, the previous pattern mode is used. This caused problems
when we needed to subtract the alight-slack for the previous stop arrival. This is originally a
optimization, but the SlackProvider handshake is error prune and should be removed.