refactor: Lookup board/alight slack by pattern
Earlier the Raptor internal SlackProvider used a handshake to retrieve the correct board/alight
slack. The worker would first set the pattern on the slack provider and then retrieve the slack.
This is unnecessary complicated and the slack can be cashed in the worker to avoid the extra
overhead to get the slack from the provider. The SlackProviderAdaptor is also simplified and two
concrete implementation is provided (forward and reverse direction).