feat: Extra stop priority cost in optimized transfer service
This adds an extra board- and alight-cost for (none) prioritized stops. A 'stopBoardAlightCosts'
is added to the generalized-cost during routing. But,this cost cannot be too high, because that would
add extra cost to the transfer, and favor other alternative paths. But, when optimizing transfers,
we do not have to take other paths into consideration and can "boost" the stop-priority-cost to
allow transfers to take place at a preferred stop.