refactor: Decouple the request and the itinerary-filter-chain
The filter chain is responsible for sorting and cropping the itinerary list is the maxNumOfItineraries is
reached. Before this commit these features were tied to the request#arriveBy. By removing this coupling
we can later use the sort-order and cropping independently of what the arriveBy is, witch is needed for
pageing to work properly.