feat: Turn ParallelRouting off
Under high load the parallel routing may lead to an entire cluster of OTP servers to go down.
We believe this happens because new connections are accepted, but the request threads block
each other, waiting for available threads to fork out. The number of accepting threads in
the Grizzly Server is 1.25x as many as cores. This is previously tested to perform well
and give the best throughput for a synchronous request handler. When switching to a
ParallelRouting strategy we need to retest this.