feat: Add support for multiple access/egress legs to the same stop
This allows us to reach the same stop on foot, car bicycle of flex in the same search. This was
not supported by the standard Raptor search before. Before the fastest access was used, but
creates problems since wa also need to find the journey with the feast transfers. This is used
in to compute the Raptor heuristics.