Initial Dockerization attempt (#46)
* Initial Dockerization attempt

This uses a multi-stage build: one for building `onebusaway-gtfs-modules`
and OpenTripPlanner, and another which is the actual image we use to run
it.

To build both of these, we're using Java binaries provided by the Adoptium
project (https://adoptium.net/), which is the rebrand of the AdoptOpenJDK
project. We currently have the AdoptOpenJDK versions listed in the
`.tool-versions` file, but they actually are no longer available.

Also of note: `onebusaway-gtfs-modules` builds properly with JDK17, but
unfortunately our version of OpenTripPlanner does not. It does build with
JDK8, so we fetch both versions.

We also pull a slightly older version of Maven, since the latest version
of Maven defaults to disallowing `http://` repositories. There may be some
quick fixes we can do to the `pom.xml` file which will allow us to use the
latest version.

The actual build steps are mostly copied verbatim from this repo's
`README`, minus the stuff about `asdf` (since as noted before, the versions
we specify there no longer exist). First, we clone and build
`onebusaway-gtfs-modules`, and then we copy the files from the local
`OpenTripPlanner` directory into the container and build that.

Once `OpenTripPlanner` is built, we begin a new stage that will become
the image we ultimately deploy. We fetch the Adoptium JRE8 (not JDK), and
then copy the necessary files from the build container (NOTE: the files we
copy are the same ones listed in `make_deploy.sh`).

The resulting image has a default `ENTRYPOINT` that will run
`OpenTripPlanner` (using the same command as in `Procfile`). The image
exposes port `5000`, which can be mapped to the appropriate port on the
host, serving plain HTTP traffic.

On my machine, the resulting image is 771MB. Around ~80MB comes from the
`debian:stable-slim` image that we're using for the base, around ~100MB
comes from the JRE, and the rest is the `OpenTripPlanner` application +
data. It may be possible to find a slimmer image for the base, I just went
with Debian because that's a distribution that I'm familiar with and is
typically fairly stable.

As far as testing this goes, I was able to point my local `dotcom` instance
at the port on my local machine, and a couple of trips I checked using the
Trip Planner functionality on the website seemed to match what I see on
prod, however I'm not super familiar with what is using OTP so it's worth
doing some more testing.

Open questions:
- Do we want to expose the plain HTTP port from the Docker image to the
  host, or should we expose the HTTPS port?
- Is there a slimmer / better base image for the final image?

* Shallow clone

* Fetch JRE in builder

* Use COPY instead of ADD for local files

* Use non-root user

* Use dumb-init as PID 1