fix: Transmodel API transport mode not present or null is all transport modes.
The default behaviour in the Sandbox Transmodel API for setting mode in the routing request is
wrong/not according to the documentation. When not present of set to 'null' the transport mode
should include all transport modes. The current behavior is an empty list witch result in
no trip plans found.