refactor: Add generic types to TransitEntity and its builder
This PR creates a new TransitEntity2 witch will replace the old TransitEntity when all
subtypes are migrated. It is a bit complex right now, but it will become simpler when
all classes are migrated. This refactoring will make real-time updates simpler, because
it allows for type-safe helper methods in the builder, modification detection, and
injecting listeners and the Deduplicator into builders at a later stage.