refactor: Remove `transitFactorIndex` and `wheelchairBoarding` from RaptorTripSchedule
The generalized cost calculation is not a responsibility of Raptor, only the Transit model and the
cost calculator implementations need to know about the properties of the trip. I have refactored so
these properties are not handled within Raptor anymore. This should make it easier to implement
different cost calculation strategies. Instead of passing the filed, the trip-schedule is passed to
the calculators.