refactor: Restrict API Resource assess by introducing OtpServerContext
The new OtpServerContext define the needed interface for all Resources. This is the first step to
remove the Router class. IntelliJ was used to extract the interface and to inline the getRouter()
method in OTPServer. I only went over the code to make sure JavaDoc and names where correctly
updated.