refactor: Move logFrequency, maxSnapshotFrequency, purgeExpiredData to timetableUpdates in config
These parameters are not updater parameters and configure the realtime updates in general. Currently,
OTP can run with either a Siri or a GTFS SnapshotProvider. These parameters are passed to the
provider. This enables us to inject the correct provider into the injector instead of duplicating
the snapshot provider creation logic in every updater.