refactor: Move Updater setup content into constructor.
In this refactoring all initialization is moved into the Updater constructor. Only used dependencies
are injected, as specific as needed. Example, the VehicleParkingUpdater does not need Graph and
TransitModel, it needs a VertexLinker and VehicleParkingService, so those are injected.

This make it possible to delete some updater fields(only needed for initialization) and to make
some fields final.