fix: forBoarding and forAlighting should be false when PickDrop is CANCELLED
This is essentially a partial revert of the following commit: https://github.com/opentripplanner/OpenTripPlanner/commit/0afeae15b3a68ff94f49e30733ef5fffb1febdd0. In particular, forBoarding and forAlighting were forced to be true for cancelled calls. The result is that the values do not reflect the planned data (despite what the documentation said) nor the realtime data (which we would expect from the data in estimated calls).