First step of untangling the inner-class 'Handler' from the containing OpenStreetMapModule.
Make Handler static, which means that it looses access to the fields in OSMModule.
Solve this by a mix of passing copies of private values to a new instance, as well
as funneling access to the data using Suppliers.

A number of tests needs to adapt to the new extended contructor.