feature: Extend with debug-itinerary-filter with 2 new profiles
The itinerary-filter-chain debug feature was just a on/off(boolean), but this is
problematic when a large amount of itineraries is deleted. To help with this I
added to new profiles LIMIT_TO_SEARCH_WINDOW and LIMIT_TO_NUM_OF_ITINERARIES.
This two profiles will list a mixture of deleted and actual itineraries, but
limit the list using the existing `search-window` or `numItineraries`.