Repair dead end barriers, which break routing
Osm contains nodes tagged as access=no and which represent an end point of a way.
For example, a doorway leading to a locked gate is sometimes modeled like that.
If routing request snaps to such end point, routing fails. Barrier end vertices do not
add any kind of useful functionality but just cause trouble.