Creates the NumItinerariesFilter and the associated NumItinerariesFilterResults to limit itineraries based on the numItineraries parameter.
Pulling this functionality into a separate filter from the MaxLimitFilter because of the need to keep more information about the filtered itineraries than just the first one that is removed. The NumItinerariesFilterResults contains information used in the page cursors currently as well as information that will be used in the future deduplication feature.