improve: Always commit the first update after a long break.
In a busy updater it does not matter if we commit early or late, as long as we
throttle the commit to happen with a given frequency. But, when testing OTP
manually, it is easy to do mistakes and assume that an update is applied when you
see "INFO 1 of 1 update messages were applied successfully (success rate: 100.0%)".
In OTP before this commit, the update was not committed, waiting to fill up the
"buffer", and if not other update was applied - nothing happened.