Fix: Allow early creation of StopModelBuilder (fixes #5497)
The problem with the current code is that the stop model builders are created at the same
time as the GraphBuilderModules. When the models are merged back the StopModel change, and
only the first GraphBuilderModule is allowed to merge its result - this happens, because a
stop model is only allowed to be merged with the model it originated from.

To solve this, we can create a stop model builder late to allow other graph builder models
to complete their work before we continue. Or, we can change the stop model builder and
stop model merge so more than one builder can exist at once. This commit take the second
approach.