refactor: Rename generalizedCost to c1 in Raptor
We use the term generalizedCost in OTP, but not in Raptor. With the new c2 value, renaming
to c1 inside Raptor make the raptor code more consistent. It also becomes easier to keep
raptor cost apart from the OTP model cost - they are slightly different and the context
is important.