feature: Make base transit-priority-group a group by its own.
There are two ways we can treat the base (local-traffic) transit priority group.
We can treat is a separate group, equal to all other groups in behaviour - or we
can treat is special and not add it to the comparason vector. Let B be the base
and G be concreate group. Then: (B) dominates (G), (G) dominates (B), (B)
dominates (BG), but (G) does not dominate (BG). With other words, paths with only
agency X(group G) is not given an advantige in the routing over paths with a
combination of agency X(group G) and local trafic (group B). This commit turn
this advanced feature OFF for now.