Fix handling of implicit access and egress mode parameters.
If the 'modes' section is present but the nested 'access' or 'egress' parameters
are not, then only transit that originates from/arrives directly at the stop should
be used for the trip, just as if the parameter was present and set to null.

Fix this by setting the corresponding mode to NOT_SET in the cases where 'modes'
is specified but the parameter is not.

Adapt and add tests to cover these cases better.