refactor: Encapsulate ordering trip-times in Timetable
All trip-times in a Timetable should be ordered by the first departure-time for the first
stop in the trip pattern. This was not enforced in the Timetable <-> TimetableBuilder, but
relied on the "outside" to do it. This PR enforces this and provides two methods for adding
trip-times to a Timetable (add & addOrUpdate).