Doc/features-explained: document GTFS feedId assignation
GTFS feedId is a fundamental concept and is an important feature to
configure properly static and realtime data together. However, this concept
is not largely explained and the information is scattered as breadcrumbs
here and there through the official documentation and reported issues.

As proposed by Leonard Ehrenfried, this commit adds a dedicated page under
"features-explained".

This fixes https://github.com/opentripplanner/OpenTripPlanner/issues/6709

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>